pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enough-journal"
version = "0.1.0"
description = "A mindful journaling tool for self-reflection and personal growth"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Stephen Piper", email = "sipos.istvan.a@gmail.com" }
]
keywords = ["journal", "mindfulness", "self-reflection", "personal-growth", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: News/Diary",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "pathlib>=1.0",
    "typing-extensions>=4.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/enough"
Issues = "https://github.com/yourusername/enough/issues"

[project.scripts]
enough = "enough.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["Core/enough"]

[tool.hatch.build]
include = [
    "Core/enough/**/*.py",
    "Core/enough/**/*.yaml",
    "Core/enough/templates/*.yaml",
    "README.md",
    "LICENSE",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "**/.DS_Store",
    "tests/**/*",
] 